Safari-611.1.10/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/idlharness.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: MediaKeyMessageEventInit inherits EventInit, but EventInit is undefined.
Description

This test verifies that implementations of the Encrypted Media Extensions API match its WebIDL definition.


Harness Error (FAIL), message = Unhandled rejection: MediaKeyMessageEventInit inherits EventInit, but EventInit is undefined.

PASS Partial interface Navigator: original interface defined
PASS Partial interface Navigator: valid exposure set
PASS Partial interface Navigator: member names are unique
PASS Partial interface HTMLMediaElement: original interface defined
PASS Partial interface HTMLMediaElement: valid exposure set
PASS Partial interface HTMLMediaElement: member names are unique
